Move post mutation helpers to hooks files
